+++ lib/src/widgets/overlay.dart
@@ -7,6 +7,8 @@ import 'package:feature_discovery/src/widgets.dart';
 import 'package:flutter/material.dart';
 
 class DescribedFeatureOverlay extends StatefulWidget {
+  static const double kDefaultBackgroundOpacity = 0.96;
+
   /// This id should be unique among all the [DescribedFeatureOverlay] widgets.
   /// Otherwise, multiple overlays would show at once, which is currently
   /// only possible if [allowShowingDuplicate] is set to `true`.
@@ -27,6 +29,10 @@ class DescribedFeatureOverlay extends StatefulWidget {
   /// The color of the large circle, where the text sits on.
   /// If null, defaults to [ThemeData.primaryColor].
   final Color backgroundColor;
+
+  /// The opacity of the large circle, where the text sits on.
+  /// If null, defaults to 0.96.
+  final double backgroundOpacity;
 
   /// Color of the target, that is the small circle behind the tap target.
   final Color targetColor;
@@ -114,7 +120,7 @@ class DescribedFeatureOverlay extends StatefulWidget {
 
   /// Duration for overlay dismiss animation.
   final Duration dismissDuration;
-  
+
   /// Controls whether the overlay should be dismissed on touching outside or not.
   ///
   /// The default value for [barrierDismissible] is `true`.
@@ -137,6 +143,7 @@ class DescribedFeatureOverlay extends StatefulWidget {
     this.enablePulsingAnimation = true,
     this.allowShowingDuplicate = false,
     this.overflowMode = OverflowMode.ignore,
+    this.backgroundOpacity = kDefaultBackgroundOpacity,
     this.openDuration = const Duration(milliseconds: 250),
     this.pulseDuration = const Duration(milliseconds: 1000),
     this.completeDuration = const Duration(milliseconds: 250),
@@ -305,11 +312,15 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay>
   }
 
   void _initAnimationControllers() {
-    _openController = AnimationController(vsync: this, duration: widget.openDuration)
-      ..addListener(() => setState(() => _transitionProgress = _openController.value));
+    _openController = AnimationController(
+        vsync: this, duration: widget.openDuration)
+      ..addListener(
+          () => setState(() => _transitionProgress = _openController.value));
 
-    _pulseController = AnimationController(vsync: this, duration: widget.pulseDuration)
-      ..addListener(() => setState(() => _transitionProgress = _pulseController.value))
+    _pulseController = AnimationController(
+        vsync: this, duration: widget.pulseDuration)
+      ..addListener(
+          () => setState(() => _transitionProgress = _pulseController.value))
       ..addStatusListener(
         (AnimationStatus status) {
           if (status == AnimationStatus.completed) {
@@ -318,10 +329,10 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay>
         },
       );
 
-    _completeController = AnimationController(
-        vsync: this, duration: widget.completeDuration)
-      ..addListener(() =>
-          setState(() => _transitionProgress = _completeController.value));
+    _completeController =
+        AnimationController(vsync: this, duration: widget.completeDuration)
+          ..addListener(() =>
+              setState(() => _transitionProgress = _completeController.value));
 
     _dismissController = AnimationController(
         vsync: this, duration: widget.dismissDuration)
@@ -637,6 +648,7 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay>
               child: _Background(
                 transitionProgress: _transitionProgress,
                 color: widget.backgroundColor ?? Theme.of(context).primaryColor,
+                defaultOpacity: widget.backgroundOpacity,
                 state: _state,
                 overflowMode: widget.overflowMode,
               ),
@@ -689,14 +701,16 @@ class _Background extends StatelessWidget {
   final double transitionProgress;
   final Color color;
   final OverflowMode overflowMode;
+  final double defaultOpacity;
 
-  const _Background({
-    Key key,
-    @required this.color,
-    @required this.state,
-    @required this.transitionProgress,
-    @required this.overflowMode,
-  })  : assert(color != null),
+  const _Background(
+      {Key key,
+      @required this.color,
+      @required this.state,
+      @required this.transitionProgress,
+      @required this.overflowMode,
+      @required this.defaultOpacity})
+      : assert(color != null),
         assert(state != null),
         assert(transitionProgress != null),
         super(key: key);
@@ -706,19 +720,19 @@ class _Background extends StatelessWidget {
       case FeatureOverlayState.opening:
         final adjustedPercent = const Interval(0.0, 0.3, curve: Curves.easeOut)
             .transform(transitionProgress);
-        return 0.96 * adjustedPercent;
+        return defaultOpacity * adjustedPercent;
 
       case FeatureOverlayState.completing:
         final adjustedPercent = const Interval(0.1, 0.6, curve: Curves.easeOut)
             .transform(transitionProgress);
 
-        return 0.96 * (1 - adjustedPercent);
+        return defaultOpacity * (1 - adjustedPercent);
       case FeatureOverlayState.dismissing:
         final adjustedPercent = const Interval(0.2, 1.0, curve: Curves.easeOut)
             .transform(transitionProgress);
-        return 0.96 * (1 - adjustedPercent);
+        return defaultOpacity * (1 - adjustedPercent);
       case FeatureOverlayState.opened:
-        return 0.96;
+        return defaultOpacity;
       case FeatureOverlayState.closed:
         return 0;
     }